# api/openapi.yaml
openapi: 3.1.0
info:
  title: Simple-VPN-Builder Control Plane API
  version: 0.1.0v
  description: High-performance VPN orchestration and client protocol provisioning control plane.
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /admins:
    get:
      tags: [Admins]
      summary: List administrators
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of admins
    post:
      tags: [Admins]
      summary: Create administrator
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Admin created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...